requirements.txt
streamlit
langchain
langchain-community
langchain-openai
openai
faiss-cpu
python-dotenv
